LC-minimum-equal-sum-of-two-arrays-after-replacing-zeros [Runtime: N/A; Memory: N/A]

+++ array/minimum-equal-sum-of-two-arrays-after-replacing-zeros.cpp
@@ -29,6 +29,6 @@ public:
             return -1;
         
         
-        return max(minSumNum1,minSumNum2);
+        return (long long)max(minSumNum1,minSumNum2);
     }
 };
